+++ lib/datalocal_for_firestore_extension.dart
@@ -1,3 +1,0 @@
-export 'src/extensions/list_data_item.dart';
-export 'src/extensions/data_item.dart';
-export 'package:datalocal/src/models/data_compute.dart';