+++ lib/datalocal_for_firestore_query_extension.dart
@@ -1,6 +1,0 @@
-export 'src/extensions/data_local.dart';
-export 'package:datalocal/src/models/data_row.dart';
-export 'src/extensions/list_data_item.dart';
-export 'src/extensions/data_row.dart';
-
-// export 'src/extensions/data_item.dart';